# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "EntityDetail": {
        "description": "Detailed information about an extracted entity.",
        "properties": {
          "end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "End position in the text. Absent when the document text was not retained, since an offset means nothing without the text.",
            "title": "End"
          },
          "label": {
            "description": "The entity label/type from the NER model",
            "title": "Label",
            "type": "string"
          },
          "score": {
            "description": "Confidence score (0-1)",
            "title": "Score",
            "type": "number"
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Start position in the text. Absent when the document text was not retained, since an offset means nothing without the text.",
            "title": "Start"
          },
          "text": {
            "description": "The extracted entity text",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "label",
          "score"
        ],
        "title": "EntityDetail",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "description": "The body of a refusal.\n\nRoutes raise HTTPException with a fixed message, and add context only when\nthe context is about the request rather than the document: the allowed file\ntypes, a size limit, a batch cap. Nothing here quotes document content.",
        "properties": {
          "allowed_extensions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "File extensions this endpoint accepts",
            "title": "Allowed Extensions"
          },
          "allowed_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "File types this endpoint accepts",
            "title": "Allowed Types"
          },
          "max_files": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Largest accepted number of files in one batch",
            "title": "Max Files"
          },
          "max_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Largest accepted file, in bytes",
            "title": "Max Size Bytes"
          },
          "message": {
            "description": "Content-free description of the failure",
            "title": "Message",
            "type": "string"
          },
          "received_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Size of the rejected file, in bytes",
            "title": "Received Size Bytes"
          },
          "received_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The rejected content type, as the client declared it",
            "title": "Received Type"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "How every refusal reaches the client.\n\nFastAPI wraps a raised HTTPException as {\"detail\": ...}, so this is the\nshape the frontend actually parses. Documenting the inner payload alone\ndescribed a body the API never sends.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExtractedEntities": {
        "description": "All extracted medical entities grouped by category.",
        "properties": {
          "anatomy": {
            "default": [],
            "description": "Anatomical structures",
            "items": {
              "$ref": "#/components/schemas/EntityDetail"
            },
            "title": "Anatomy",
            "type": "array"
          },
          "examinations": {
            "default": [],
            "description": "Medical examinations",
            "items": {
              "$ref": "#/components/schemas/EntityDetail"
            },
            "title": "Examinations",
            "type": "array"
          },
          "measurements": {
            "default": [],
            "description": "Measurements and values",
            "items": {
              "$ref": "#/components/schemas/EntityDetail"
            },
            "title": "Measurements",
            "type": "array"
          },
          "other": {
            "default": [],
            "description": "Other medical entities",
            "items": {
              "$ref": "#/components/schemas/EntityDetail"
            },
            "title": "Other",
            "type": "array"
          },
          "pathologies": {
            "default": [],
            "description": "Diseases and conditions",
            "items": {
              "$ref": "#/components/schemas/EntityDetail"
            },
            "title": "Pathologies",
            "type": "array"
          },
          "patient_info": {
            "default": [],
            "description": "Patient demographic information",
            "items": {
              "$ref": "#/components/schemas/EntityDetail"
            },
            "title": "Patient Info",
            "type": "array"
          },
          "symptoms": {
            "default": [],
            "description": "Signs and symptoms",
            "items": {
              "$ref": "#/components/schemas/EntityDetail"
            },
            "title": "Symptoms",
            "type": "array"
          },
          "temporal": {
            "default": [],
            "description": "Temporal information",
            "items": {
              "$ref": "#/components/schemas/EntityDetail"
            },
            "title": "Temporal",
            "type": "array"
          },
          "treatments": {
            "default": [],
            "description": "Treatments and medications",
            "items": {
              "$ref": "#/components/schemas/EntityDetail"
            },
            "title": "Treatments",
            "type": "array"
          }
        },
        "title": "ExtractedEntities",
        "type": "object"
      },
      "ExtractionResponse": {
        "properties": {
          "expires_in_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Seconds before the stored document is automatically deleted",
            "title": "Expires In Seconds"
          },
          "extracted_entities": {
            "$ref": "#/components/schemas/ExtractedEntities",
            "description": "Medical entities found in the text with detailed information"
          },
          "file_id": {
            "description": "Unique identifier of the processed file",
            "title": "File Id",
            "type": "string"
          },
          "mapping_info": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Information about the label mapping used (language, dataset)",
            "title": "Mapping Info"
          },
          "processed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the extraction was completed",
            "title": "Processed At"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Extracted text from the document. Null unless the deployment opted into storing document text.",
            "title": "Text"
          }
        },
        "required": [
          "file_id",
          "extracted_entities"
        ],
        "title": "ExtractionResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Med-Assist Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/get_extracted_text/{file_id}": {
      "get": {
        "description": "Endpoint to retrieve the entities extracted from a processed file.\n\nEntities are extracted at upload time, so this only reads what was stored.\nThe document text comes back only when the deployment kept it.\n\nArgs:\n    file_id: The unique identifier of the file to read\n    text_repository: Injected text repository dependency\n    entity_extractor: Injected entity extractor dependency\n\nReturns:\n    ExtractionResponse: Contains the file ID, identified medical entities and,\n    when retained, the extracted text\n\nRaises:\n    HTTPException: 404 if file not found or not processed yet",
        "operationId": "get_extracted_text_api_get_extracted_text__file_id__get",
        "parameters": [
          {
            "description": "Unique identifier of the uploaded file to read entities from",
            "in": "path",
            "name": "file_id",
            "required": true,
            "schema": {
              "description": "Unique identifier of the uploaded file to read entities from",
              "examples": [
                "123e4567-e89b-12d3-a456-426614174000"
              ],
              "title": "File Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid file ID format"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "File not found or not processed yet"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The model is not loaded yet"
          }
        },
        "summary": "Get Extracted Text",
        "tags": [
          "Text Extractions"
        ]
      }
    },
...
  }
}